refactor(idp-basic): Remove duplicate IdP legacy garbage collector
Drop IdpLegacyGarbageCollector and IdpStorageBootstrap now that GC runs
with IdpRelationalEntityStore via IdpRelationalGarbageCollector.
